% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_copy_table_to_db.R
\name{db_copy_table_to_db}
\alias{db_copy_table_to_db}
\title{db_copy_table_to_db}
\usage{
db_copy_table_to_db(table_name, con_from, con_to, batch_size = 10000)
}
\arguments{
\item{table_name}{name of table}

\item{con_from}{connection to copy table from}

\item{con_to}{connection to copy table to}

\item{batch_size}{how many lines should be copied at once}
}
\description{
db_copy_table_to_db
}
