% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlValues.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{sqlValues,DBIConnection,vector-method}
\alias{sqlValues,DBIConnection,data.frame-method}
\title{hidden aliases}
\usage{
\S4method{sqlValues}{DBIConnection,vector}(conn, x, ..., val_sep = "\\n")

\S4method{sqlValues}{DBIConnection,data.frame}(conn, x, ..., val_sep = "\\n")
}
\description{
hidden aliases
}
