% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_operator_sql.R
\name{+.SQL}
\alias{+.SQL}
\title{+-Operator for SQL}
\usage{
\method{+}{SQL}(a, b)
}
\arguments{
\item{a}{first part of SQL statement}

\item{b}{second part of SQL statement}
}
\description{
+-Operator for SQL
}
