% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_check_job_execution_necessary.R
\name{wpd_check_job_execution_necessary}
\alias{wpd_check_job_execution_necessary}
\title{wpd_check_job_execution_necessary}
\usage{
wpd_check_job_execution_necessary(date, lang)
}
\arguments{
\item{date}{task_date}

\item{lang}{task_lang}
}
\description{
wpd_check_job_execution_necessary
}
