% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_create_tables.R
\name{wpd_create_tables}
\alias{wpd_create_tables}
\title{wpd_create_tables}
\usage{
wpd_create_tables(flatten = TRUE, verbose = TRUE, node = NULL)
}
\arguments{
\item{flatten}{flatten query results}

\item{verbose}{be talkative?}
}
\description{
wpd_create_tables
}
