% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_download_month.R
\name{wpd_download_month}
\alias{wpd_download_month}
\title{wpd_download_month}
\usage{
wpd_download_month(year, month)
}
\arguments{
\item{year}{year for whic hto download data}

\item{month}{month for which to download data}
}
\value{
nothing, called for side effects
}
\description{
wpd_download_month
}
