% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_dump_lines_to_df.R
\name{wpd_dump_lines_to_df_list}
\alias{wpd_dump_lines_to_df_list}
\title{wpd_dump_lines_to_df_list}
\usage{
wpd_dump_lines_to_df_list(lines, filter = TRUE, split = TRUE)
}
\arguments{
\item{lines}{character vector of lines from a wikipedia üage views dump file
should be filtered beforehand to exclude}
}
\description{
wpd_dump_lines_to_df_list
}
