% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_filter_extract.R
\name{wpd_filter_extract}
\alias{wpd_filter_extract}
\title{wpd_filter_extract}
\usage{
wpd_filter_extract(file, languages = wpd_languages)
}
\arguments{
\item{fname}{file name to extract data from}

\item{lang}{language to extract}
}
\description{
wpd_filter_extract
}
