% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_filter_extract_worker.R
\name{wpd_filter_extract_worker}
\alias{wpd_filter_extract_worker}
\title{wpd_filter_extract_worker}
\usage{
wpd_filter_extract_worker(fname, lang)
}
\arguments{
\item{fname}{file name to filter extract}

\item{lang}{language to extract}
}
\description{
wpd_filter_extract_worker
}
