% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_get_dump_links.R
\name{wpd_get_dump_links}
\alias{wpd_get_dump_links}
\title{get links for dump downloads}
\usage{
wpd_get_dump_links(out_file = tempfile(pattern = "dump_links", fileext =
  ".txt"))
}
\arguments{
\item{directory}{path to store dump_links.txt file}

\item{force}{force re-downloading and re-extracting dump links}
}
\description{
get links for dump downloads
}
