% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_get_dumps.R
\name{wpd_get_dumps}
\alias{wpd_get_dumps}
\title{function that downloads dump files for specific dates}
\usage{
wpd_get_dumps(ts, directory = ".")
}
\arguments{
\item{ts}{timestamp e.g. "20071201" or shorter that will be used to match dump links}

\item{directory}{directory to store downloads in}
}
\description{
function that downloads dump files for specific dates
}
