% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_get_queries.R
\name{wpd_get_queries}
\alias{wpd_get_queries}
\title{wpd_get_query}
\usage{
wpd_get_queries(queries, verbose = TRUE, con = NULL, flatten = TRUE)
}
\arguments{
\item{verbose}{return extra information -- timings, potential database errors -- or not}

\item{query}{SQL statement to execute}
}
\description{
Establish a connection, execute SQL statement against database, close connection and return.
}
\examples{


}
