% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_get_time_series.R
\name{wpd_get_time_series}
\alias{wpd_get_time_series}
\title{wpd_get_time_series}
\usage{
wpd_get_time_series(page = NULL, lang = NULL)
}
\arguments{
\item{page}{article name}

\item{lang}{article language}
}
\description{
wpd_get_time_series
}
