% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_hlp_subset_dots.R
\name{wpd_hlp_subset_dots}
\alias{wpd_hlp_subset_dots}
\title{wpd_hlp_subset_dots}
\usage{
wpd_hlp_subset_dots(..., i = NULL)
}
\arguments{
\item{...}{parameter to subset}

\item{i}{index to subset by}
}
\value{
List of its inputs where inputs are each subset according to i
}
\description{
wpd_hlp_subset_dots
}
