% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_job_log.R
\name{wpd_job_log}
\alias{wpd_job_log}
\title{Title}
\usage{
wpd_job_log(date, lang, file, start_time, progress, job_id)
}
\arguments{
\item{date}{date covered by job}

\item{lang}{language covered by job}

\item{file}{input file}

\item{start_time}{start time}

\item{progress}{some measure of progress}
}
\description{
Title
}
