% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_job_new.R
\name{wpd_job_new}
\alias{wpd_job_new}
\title{wpd_job_new}
\usage{
wpd_job_new(lang, date, file, job_type)
}
\arguments{
\item{lang}{language covered by job}

\item{date}{date covered by job}

\item{file}{input file}
}
\description{
wpd_job_new
}
