% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_job_update.R
\name{wpd_job_update}
\alias{wpd_job_update}
\title{wpd_job_update}
\usage{
wpd_job_update(job_id, ...)
}
\arguments{
\item{job_id}{the job id to change}

\item{...}{a number of key value pairs}
}
\description{
wpd_job_update
}
