% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_nodes.R
\docType{data}
\name{wpd_nodes}
\alias{wpd_nodes}
\title{list of server nodes}
\format{An object of class \code{character} of length 2.}
\usage{
wpd_nodes
}
\description{
list of server nodes
}
\keyword{datasets}
