% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_nodes_master.R
\docType{data}
\name{wpd_nodes_master}
\alias{wpd_nodes_master}
\title{name and host of master node}
\format{An object of class \code{character} of length 1.}
\usage{
wpd_nodes_master
}
\description{
name and host of master node
}
\keyword{datasets}
