% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_push_dict_helper_sql_insert.R
\name{wpd_push_dict_helper_sql_insert}
\alias{wpd_push_dict_helper_sql_insert}
\title{wpd_push_dict_helper_sql_insert}
\usage{
wpd_push_dict_helper_sql_insert(pages, lang, con = NULL)
}
\arguments{
\item{page}{character vector of page names}
}
\description{
wpd_push_dict_helper_sql_insert
}
