% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_push_item.R
\name{wpd_push_item}
\alias{wpd_push_item}
\title{wpd_push_item}
\usage{
wpd_push_item(date, lang, page, views, con = NULL)
}
\arguments{
\item{date}{date for which to push}

\item{lang}{language}

\item{page}{page names}

\item{views}{page views}

\item{con}{connection object}
}
\description{
wpd_push_item
}
