% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_sql.R
\name{wpd_sql}
\alias{wpd_sql}
\title{wpd_sql}
\usage{
wpd_sql(query_string, ...)
}
\arguments{
\item{query_string}{a query string}

\item{...}{values to put into string, see \link{sprintf}}
}
\description{
wpd_sql
}
