% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_sql_glue.R
\name{wpd_sql_glue}
\alias{wpd_sql_glue}
\title{wpd_sql_glue}
\usage{
wpd_sql_glue(query_string, ..., .con = DBI::ANSI(),
  .envir = parent.frame(), .na = DBI::SQL("NULL"))
}
\arguments{
\item{query_string}{a query string}

\item{...}{values to put into string, see \link{glue::glue_sql}}
}
\description{
wpd_sql_glue
}
