% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_task_update.R
\name{wpd_task_update}
\alias{wpd_task_update}
\title{wpd_task_update}
\usage{
wpd_task_update(task_id, ...)
}
\arguments{
\item{task_id}{the job id to change}

\item{...}{a number of key value pairs}
}
\description{
wpd_task_update
}
