% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_upload_new_job.R
\name{wpd_upload_new_job}
\alias{wpd_upload_new_job}
\title{wpd_upload_new_job}
\usage{
wpd_upload_new_job(lang, date, file)
}
\arguments{
\item{lang}{language covered by job}

\item{date}{date covered by job}

\item{file}{input file}
}
\description{
wpd_upload_new_job
}
