% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_upload_pageview_counts.R
\name{wpd_upload_pageview_counts}
\alias{wpd_upload_pageview_counts}
\title{wpd_upload_pageview_counts}
\usage{
wpd_upload_pageview_counts(page_name, page_view_count, page_view_date,
  page_language, conn = NULL, upload_type = c("bz2", "gz"))
}
\arguments{
\item{page_name}{character vector of page names}

\item{page_view_count}{numeric vector of page view counts}

\item{page_view_date}{date for page view counts (length 1)}

\item{page_language}{language (length 1)}

\item{conn}{either NULL or an connection object}
}
\description{
wpd_upload_pageview_counts
}
