% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpd_dispatcher.R
\name{wpd_dispatcher}
\alias{wpd_dispatcher}
\title{wpd_dispatcher}
\usage{
wpd_dispatcher()
}
\description{
function that can be called by cron job via Rscript -e "wpd::wpd_dispatcher()"
to dispatch scripts
}
